+++ CHANGES.md
@@ -4,7 +4,7 @@ Versions
 1.36.1 (2026-05-26)
 -------------------
 
-- Fix to dateutil.local_to_utc to take both datetime
+- Fix to banzai.utils.date_utils.local_to_utc to take both datetime
   and time objects.
 
 1.36.0 (2026-05-08)